requirements.txt
aiohttp==3.7.4.post0
APScheduler==3.7.0
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
chardet==4.0.0
chromedriver-autoinstaller==0.2.2
click==7.1.2
expand-string==0.0.2
Flask==1.1.2
idna==3.1
itsdangerous==1.1.0
Jinja2==2.11.3
lxml==4.6.3
MarkupSafe==1.1.1
multidict==5.1.0
python-dotenv==0.17.0
pytz==2021.1
selenium==3.141.0
six==1.15.0
slack==0.0.2
slackclient==2.9.3
soupsieve==2.2.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.4
Werkzeug==1.0.1
yarl==1.6.3
